code/requirements.txt
plaid-python>=5.0.0, < 6.0.0
